Do not ignore mansync in display-x. If xliblock is held while mansync tries to flush, the flush is ignored. That can potentially lead to no flushes at all, which is not very polite.
Further, in order to fix display-x to not call X functions
directly from mansync (which is in signal handler
context if sigsched is used) the flush has to be delayed
somehow, this patch solves that issue by using a new
lock, named flushlock.

flushlock is checked every time xliblock is unlocked, which
happen regularly in input-xwin. So, in the uncontended case
the actual flush happens from input-xwin and in the
contended case the flush happens before returning to the
app from whatever call caused xliblock to be locked when
mansync wanted to flush.

The function pointer priv->lock_xliblock (and the macro
GGI_X_LOCK_XLIB) is strictly speaking not needed, but I
added it for symmetry.