Permit kgiGetResource() to return an error code
Unlike all the other functions in kgi.c, kgiGetResource() was returning
the actual result structure as the function return value, instead of a
status code. Thus it was limited to returning NULL as an indicator that
something failed, without any further details.

Stop this discrimination by changing it to pass the result through a
pointer argument instead, and returning error codes like everyone else.